Updated docker scripts

+++ Docker/preparation.sh
@@ -30,9 +30,12 @@ if [ ! -f ./${output}/4_bam_preparation/resources-broad-hg38-v0-Homo_sapiens_ass
      mv Homo_sapiens_assembly38.dict ./${output}/4_bam_preparation/resources-broad-hg38-v0-Homo_sapiens_assembly38.dict && \
      echo "Locate "${output}"/4_bam_preparation/resources-broad-hg38-v0-Homo_sapiens_assembly38.dict";
 fi
-if [ ! -f ./${output}/4_bam_preparation/resources-broad-hg38-v0-Homo_sapiens_assembly38.fasta ]; then
+if [ ! -f ./Homo_sapiens_assembly38.fasta ]; then
      curl -OL https://storage.googleapis.com/genomics-public-data/resources/broad/hg38/v0/Homo_sapiens_assembly38.fasta;
-     mv Homo_sapiens_assembly38.fasta ./${output}/4_bam_preparation/resources-broad-hg38-v0-Homo_sapiens_assembly38.fasta && \
+     echo "Locate Homo_sapiens_assembly38.fasta";
+fi
+if [ ! -f ./${output}/4_bam_preparation/resources-broad-hg38-v0-Homo_sapiens_assembly38.fasta ]; then
+     cp Homo_sapiens_assembly38.fasta ./${output}/4_bam_preparation/resources-broad-hg38-v0-Homo_sapiens_assembly38.fasta && \
      echo "Locate "${output}"/4_bam_preparation/resources-broad-hg38-v0-Homo_sapiens_assembly38.fasta";
 fi
 if [ ! -f ./${output}/4_bam_preparation/resources-broad-hg38-v0-Homo_sapiens_assembly38.fasta.fai ]; then
@@ -105,24 +108,33 @@ fi
 
 if [ ! -f ${output}/4_bam_preparation/bam_preparation_v2.sh ]; then
 cat << EOT >> ${output}/4_bam_preparation/bam_preparation_v2.sh
-# bam_preparation.sh <reference.fa> <mapped.bam> <readname> <platform> <output.bam> <memory>
-# update read group
-echo "Update Read Group..."
-docker run \
-     -u "\$(id -u \$USER):\$(id -g \$USER)" \
-     -v /etc/passwd:/etc/passwd:ro \
-     -v /etc/group:/etc/group:ro \
-     --memory "\$6"g -itv \$PWD:/data -w /data --rm biocontainers/picard:2.3.0 \
-     picard AddOrReplaceReadGroups \
-     INPUT="\$2" \
-     OUTPUT="\$2".addRG.bam \
-     SO=coordinate \
-     RGID=FLOWCELLID \
-     RGLB="\$3"_library1 \
-     RGPU=unknown \
-     RGPL="\$4" \
-     RGSM="\$3" \
-     CREATE_INDEX=true;
+if [[ ! -f "\$2".addRG.bam ]]; then
+     # bam_preparation.sh <reference.fa> <mapped.bam> <readname> <platform> <output.bam> <memory>
+     # update read group
+     echo "Update Read Group..."
+     docker run \
+          -u "\$(id -u \$USER):\$(id -g \$USER)" \
+          -v /etc/passwd:/etc/passwd:ro \
+          -v /etc/group:/etc/group:ro \
+          --memory "\$6"g -itv \$PWD:/data -w /data --rm biocontainers/picard:2.3.0 \
+          picard AddOrReplaceReadGroups \
+          INPUT="\$2" \
+          OUTPUT="\$2".addRG.bam \
+          SO=coordinate \
+          RGID=FLOWCELLID \
+          RGLB="\$3"_library1 \
+          RGPU=unknown \
+          RGPL="\$4" \
+          RGSM="\$3" \
+          CREATE_INDEX=true;
+	echo "DONE"
+fi
+if [[ ! -f "\$2".addRG.bam ]]; then
+echo "Error."
+exit 1;
+fi
+
+if [[ ! -f "\$2".addRG.duprm.bam ]]; then
 # mark duplication
 echo "Mark duplication..."
 docker run \
@@ -134,19 +146,30 @@ docker run \
      -I "\$2".addRG.bam \
      -O "\$2".addRG.duprm.bam \
      -M MarkDuplicatesSpark_output.metrics.txt;
+	echo "DONE"
+fi
+if [[ ! -f "\$2".addRG.duprm.bam ]]; then
+    echo "Error."
+    exit 1;
+fi
 
-# Splits reads that contain Ns in their cigar string
-echo "Split reads that contain Ns in their cigar string..."
-docker run \
-     -u "\$(id -u \$USER):\$(id -g \$USER)" \
-     -v /etc/passwd:/etc/passwd:ro \
-     -v /etc/group:/etc/group:ro \
-     --memory "\$6"g -itv \$PWD:/data -w /data --rm broadinstitute/gatk:4.3.0.0 \
-     gatk SplitNCigarReads \
-     -R \$1 \
-     -I "\$2".addRG.duprm.bam \
-     -O \$5;
-echo "DONE"
+if [[ ! -f \$5 ]]; then
+     # Splits reads that contain Ns in their cigar string
+     echo "Split reads that contain Ns in their cigar string..."
+     docker run \
+          -u "\$(id -u \$USER):\$(id -g \$USER)" \
+          -v /etc/passwd:/etc/passwd:ro \
+          -v /etc/group:/etc/group:ro \
+          --memory "\$6"g -itv \$PWD:/data -w /data --rm broadinstitute/gatk:4.3.0.0 \
+          gatk SplitNCigarReads \
+          -R \$1 \
+          -I "\$2".addRG.duprm.bam \
+          -O \$5;
+     echo "DONE"
+if [[ ! -f \$5 ]]; then
+    echo "Error."
+    exit 1;
+fi
 EOT
 fi
 chmod +x ${output}/4_bam_preparation/bam_preparation_v2.sh;
